requirements.txt
httpx>=0.27.0
openai>=1.30.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
firecrawl-py>=1.0.0
